Stub elasticsearch Flow
Needs filling in; currently just getting to grips with how this should
be done with akka streams. ElasticsearchReporting will define the flow
for transforming a Report into a proper HttpRequest to elasticsearch,
and then that can be run through a HostConnectionPool, which is another
Flow, to transform it into a response.